// packages/ui/src/_generated/palette/cyan.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const cyan = stylex.defineConsts({
  _0: "#000000",
  _2: "#000810",
  _5: "#00131E",
  _7: "#001824",
  _9: "#001C2A",
  _11: "#00202F",
  _13: "#002435",
  _20: "#00344A",
  _30: "#004D6B",
  _40: "#006990",
  _50: "#0085B6",
  _60: "#1AA1D9",
  _70: "#48BCF6",
  _80: "#8CD4FF",
  _90: "#CAE9FF",
  _92: "#D5EDFF",
  _95: "#E5F3FF",
  _97: "#F0F8FF",
  _98: "#F6FAFF",
  _99: "#FBFCFF",
  _100: "#FFFFFF",
});

export const cyan_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "0,8,16",
  _5: "0,19,30",
  _7: "0,24,36",
  _9: "0,28,42",
  _11: "0,32,47",
  _13: "0,36,53",
  _20: "0,52,74",
  _30: "0,77,107",
  _40: "0,105,144",
  _50: "0,133,182",
  _60: "26,161,217",
  _70: "72,188,246",
  _80: "140,212,255",
  _90: "202,233,255",
  _92: "213,237,255",
  _95: "229,243,255",
  _97: "240,248,255",
  _98: "246,250,255",
  _99: "251,252,255",
  _100: "255,255,255",
});