// packages/ui/src/_generated/palette/green.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const green = stylex.defineConsts({
  _0: "#000000",
  _2: "#000A01",
  _5: "#001503",
  _7: "#001A05",
  _9: "#001E06",
  _11: "#002207",
  _13: "#002609",
  _20: "#003710",
  _30: "#00581E",
  _40: "#008431",
  _50: "#05B046",
  _60: "#48D766",
  _70: "#6DFA84",
  _80: "#D2FFCF",
  _90: "#EEFFE9",
  _92: "#F2FFED",
  _95: "#FBFFF6",
  _97: "#FFFFFF",
  _98: "#FFFFFF",
  _99: "#FFFFFF",
  _100: "#FFFFFF",
});

export const green_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "0,10,1",
  _5: "0,21,3",
  _7: "0,26,5",
  _9: "0,30,6",
  _11: "0,34,7",
  _13: "0,38,9",
  _20: "0,55,16",
  _30: "0,88,30",
  _40: "0,132,49",
  _50: "5,176,70",
  _60: "72,215,102",
  _70: "109,250,132",
  _80: "210,255,207",
  _90: "238,255,233",
  _92: "242,255,237",
  _95: "251,255,246",
  _97: "255,255,255",
  _98: "255,255,255",
  _99: "255,255,255",
  _100: "255,255,255",
});