// packages/ui/src/_generated/palette/orange.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const orange = stylex.defineConsts({
  _0: "#000000",
  _2: "#100500",
  _5: "#1D0C00",
  _7: "#231000",
  _9: "#291300",
  _11: "#2E1600",
  _13: "#331A00",
  _20: "#482600",
  _30: "#703F00",
  _40: "#A65F00",
  _50: "#DB7F00",
  _60: "#FF9F36",
  _70: "#FFC693",
  _80: "#FFE0C7",
  _90: "#FFEDE1",
  _92: "#FFF0E6",
  _95: "#FFF4EE",
  _97: "#FFF8F5",
  _98: "#FFFAF9",
  _99: "#FFFCFF",
  _100: "#FFFFFF",
});

export const orange_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "16,5,0",
  _5: "29,12,0",
  _7: "35,16,0",
  _9: "41,19,0",
  _11: "46,22,0",
  _13: "51,26,0",
  _20: "72,38,0",
  _30: "112,63,0",
  _40: "166,95,0",
  _50: "219,127,0",
  _60: "255,159,54",
  _70: "255,198,147",
  _80: "255,224,199",
  _90: "255,237,225",
  _92: "255,240,230",
  _95: "255,244,238",
  _97: "255,248,245",
  _98: "255,250,249",
  _99: "255,252,255",
  _100: "255,255,255",
});