// packages/ui/src/_generated/palette/pink.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const pink = stylex.defineConsts({
  _0: "#000000",
  _2: "#1F0003",
  _5: "#360009",
  _7: "#42000D",
  _9: "#4F0011",
  _11: "#5D0016",
  _13: "#6B001A",
  _20: "#9C002B",
  _30: "#CF003B",
  _40: "#E61346",
  _50: "#FC2A53",
  _60: "#FF6675",
  _70: "#FF9399",
  _80: "#FFB9BB",
  _90: "#FFDCDC",
  _92: "#FFE3E3",
  _95: "#FFEDED",
  _97: "#FFF4F4",
  _98: "#FFF8F7",
  _99: "#FFFBFF",
  _100: "#FFFFFF",
});

export const pink_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "31,0,3",
  _5: "54,0,9",
  _7: "66,0,13",
  _9: "79,0,17",
  _11: "93,0,22",
  _13: "107,0,26",
  _20: "156,0,43",
  _30: "207,0,59",
  _40: "230,19,70",
  _50: "252,42,83",
  _60: "255,102,117",
  _70: "255,147,153",
  _80: "255,185,187",
  _90: "255,220,220",
  _92: "255,227,227",
  _95: "255,237,237",
  _97: "255,244,244",
  _98: "255,248,247",
  _99: "255,251,255",
  _100: "255,255,255",
});