// packages/ui/src/_generated/palette/purple.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const purple = stylex.defineConsts({
  _0: "#000000",
  _2: "#11001C",
  _5: "#1F002F",
  _7: "#250038",
  _9: "#2A003F",
  _11: "#2F0046",
  _13: "#34004D",
  _20: "#470067",
  _30: "#6B0098",
  _40: "#8F31BE",
  _50: "#B154E0",
  _60: "#CC6DFB",
  _70: "#DA8EFF",
  _80: "#E8B2FF",
  _90: "#F6D8FF",
  _92: "#F9E0FF",
  _95: "#FDEBFF",
  _97: "#FFF3FE",
  _98: "#FFF7FC",
  _99: "#FFFBFF",
  _100: "#FFFFFF",
});

export const purple_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "17,0,28",
  _5: "31,0,47",
  _7: "37,0,56",
  _9: "42,0,63",
  _11: "47,0,70",
  _13: "52,0,77",
  _20: "71,0,103",
  _30: "107,0,152",
  _40: "143,49,190",
  _50: "177,84,224",
  _60: "204,109,251",
  _70: "218,142,255",
  _80: "232,178,255",
  _90: "246,216,255",
  _92: "249,224,255",
  _95: "253,235,255",
  _97: "255,243,254",
  _98: "255,247,252",
  _99: "255,251,255",
  _100: "255,255,255",
});