// packages/ui/src/_generated/palette/yellow.stylex.ts
// AUTO-GENERATED by packages/system-palette-codegen — do not edit.
// To regenerate: pnpm codegen:palette

import * as stylex from "@stylexjs/stylex";

export const yellow = stylex.defineConsts({
  _0: "#000000",
  _2: "#0C0700",
  _5: "#171000",
  _7: "#1D1500",
  _9: "#221900",
  _11: "#271D00",
  _13: "#2C2100",
  _20: "#3F3100",
  _30: "#624D00",
  _40: "#8D7000",
  _50: "#BA9400",
  _60: "#E5B700",
  _70: "#FFDB74",
  _80: "#FFF2D6",
  _90: "#FFF7EC",
  _92: "#FFF8F0",
  _95: "#FFFAF8",
  _97: "#FFFDFF",
  _98: "#FFFEFF",
  _99: "#FFFFFF",
  _100: "#FFFFFF",
});

export const yellow_rgb = stylex.defineConsts({
  _0: "0,0,0",
  _2: "12,7,0",
  _5: "23,16,0",
  _7: "29,21,0",
  _9: "34,25,0",
  _11: "39,29,0",
  _13: "44,33,0",
  _20: "63,49,0",
  _30: "98,77,0",
  _40: "141,112,0",
  _50: "186,148,0",
  _60: "229,183,0",
  _70: "255,219,116",
  _80: "255,242,214",
  _90: "255,247,236",
  _92: "255,248,240",
  _95: "255,250,248",
  _97: "255,253,255",
  _98: "255,254,255",
  _99: "255,255,255",
  _100: "255,255,255",
});